Add files via upload
The initial list of packages to begin for supported linux distros. And use in the next task for introducing a stu test to go through the installation and assemble the list of packages required and compare with the canonical list and build it. I will start with centos to begin.