BLDMGR-10183: Linux installer instructs to install 'xcb-util-cursor' on Rocky 8.10, but no such package exists (#34)
### Description

This PR updates the `check_required_packages.sh` script to generate
correct and more helpful installation commands for users on modern
RHEL-based systems (RHEL, CentOS, Rocky):

* Conditional Package Manager: The script now detects the major OS
version. It will use dnf in its output for systems newer than version 7
(e.g., RHEL 8/9) and continue using yum for older systems like CentOS 7.
* Recommend EPEL Installation: The script now correctly identifies that
`epel-release` is a prerequisite on RHEL-based systems and includes it
in the installation command it generates for the user.